edit histogram class etc.

+++ Floodmap.cpp
@@ -2,15 +2,12 @@
 
 Floodmap::Floodmap()
 {
-    I0 = cv::Mat::zeros(nPixel, nPixel, CV_32FC1);
+    I0 = cv::Mat::zeros(nPixel, nPixel, CV_64FC1);
     m_segr = cv::Mat::zeros(nPixel, nPixel, CV_16UC1);
 }
 
 
-Floodmap::~Floodmap()
-{
-
-}
+Floodmap::~Floodmap() { }
 
 
 void Floodmap::SetCrysatlN(int n)
@@ -29,11 +26,9 @@ cv::Mat_<quint16> Floodmap::GetSegResult()
 }
 
 
-void Floodmap::SetMap(cv::Mat_<float> I)
+void Floodmap::SetMap(cv::Mat_<qreal> I)
 {
-
     I0 = I;
-
 }
